backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
qrcode[pil]==7.4.2
pytz==2024.1
python-multipart==0.0.9
httpx>=0.27.0
gspread>=6.0.0
google-auth>=2.0.0
